chore: Fix Linux build.
The newer Bazel version gives a problem with Boost.
Pinning that temporarily; looks like upstream boost_rules
has already been updated.

+++ indexer/DriverWorkerComms.h
@@ -8,7 +8,7 @@
 
 namespace scip_clang {
 
-using WorkerId = unsigned long long;
+using WorkerId = uint64_t;
 
 std::string driverToWorkerQueueName(std::string_view driverId,
                                     WorkerId workerId);
